+++ CHANGELOG.md
@@ -10,9 +10,15 @@ and this project adheres to a _modified_ form of _[Semantic Versioning][semver]_
 
 ### Added
 
+- Added tests fot `stelae git` ([64])
+
 ### Changed
 
+- Merged `stelae git` and `stelae serve` into single command ([64])
+
 ### Fixed
+
+- git serve now support commitish that contains / in name ([64])
 
 ### Removed